+++ CHANGELOG.md
@@ -168,3 +168,4 @@ def obfuscate_query(query: str, technique_name: str, triggers: Optional[List[str
 - [2026-05-17 12:21] home/godmode/lib/encoder.py
 - [2026-05-17 12:23] home/godmode/lib/detectors.py
 - [2026-05-17 12:24] home/godmode/engine/racer.py
+- [2026-05-17 12:26] home/godmode/engine/scorer.py